将公用部分抽取出来
